Add support to track multiple users
So far, we were restricted to profile a single user. A generic
support requires that we take care of every process in the system.
This patch adds required support to track every user. The root user
is still exluded to keep the process list short. This patch should
take care of processes like KVM and redis, that operates with a
separate username.

Signed-off-by: Panwar, Ashish <panwarit014@gmail.com>